misc/samsung/scsc: Fix 'va_start' has undefined behavior [-Wvarargs]
../drivers/misc/samsung/scsc/scsc_wifilogger_ring_connectivity.c:101:15: warning: passing an object that undergoes default argument promotion to 'va_start' has undefined behavior [-Wvarargs]
        va_start(ap, tag_count);
                     ^
../drivers/misc/samsung/scsc/scsc_wifilogger_ring_connectivity.c:82:34: note: parameter of type 'u8' (aka 'unsigned char') is declared here
                                                   u16 driver_event_id, u8 tag_count, ...)
                                                                           ^

+++ drivers/misc/samsung/scsc/scsc_wifilogger_ring_connectivity.c
@@ -79,7 +79,7 @@ int scsc_wifilogger_ring_connectivity_fw_event(wlog_verbose_level lev, u16 fw_ev
 EXPORT_SYMBOL(scsc_wifilogger_ring_connectivity_fw_event);
 
 int scsc_wifilogger_ring_connectivity_driver_event(wlog_verbose_level lev,
-						   u16 driver_event_id, u8 tag_count, ...)
+						   u16 driver_event_id, int tag_count, ...)
 {
 	int							i;
 	u64							timestamp;
